Refactor refactor 'recommenders' folder to provide abstraction and simple interfaces
