+++ RecommenderModule/recommenders/item_based_example.txt
@@ -1,9 +1,9 @@
-from item_based_collaborative_filtering import ItemBasedCollaborativeFiltering
+from item_based_recommender import ItemBasedRecommender
 
 # Number of minimum ratings for a book to be kept in the filtered ratings dataset
 filtering_min_ratings_threshold = 15
 # We use item-based collaborative filtering
-item_based = ItemBasedCollaborativeFiltering(filtering_min_ratings_threshold)
+item_based = ItemBasedRecommender(filtering_min_ratings_threshold)
 
 # Test
 test_subject = 278188
